requirements.txt
# Core Flask dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# Azure AI and OpenAI dependencies
azure-ai-projects==1.0.0b4
azure-identity==1.15.0
openai==1.51.0

# Alternative Azure dependencies (fallbacks)
azure-core==1.29.6
azure-storage-blob==12.19.0
azure-cognitiveservices-language-textanalytics==5.3.0

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows

# Utility dependencies
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0  # For production deployment

# Development and debugging
flask-cors==4.0.0